+++ Superuser/jni/Android.mk
@@ -6,7 +6,7 @@ LOCAL_MODULE := su
 LOCAL_LDFLAGS := -static
 LOCAL_STATIC_LIBRARIES := sqlite3
 LOCAL_C_INCLUDES := $(LOCAL_PATH)/sqlite3
-LOCAL_SRC_FILES := su/su.c su/activity.c su/db.c su/utils.c su/daemon.c
+LOCAL_SRC_FILES := su/su.c su/activity.c su/db.c su/utils.c su/daemon.c su/pts.c
 include $(BUILD_EXECUTABLE)
